rename example, remove some junk
