--- examples/express/client/src/main.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-
-import App from "./App.tsx";
-
-import "./index.css";
-
-ReactDOM.createRoot(document.getElementById("root")!).render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-);
