--- examples/express/client/tailwind.config.js
@@ -1,10 +1,0 @@
-import { withUt } from "uploadthing/tw";
-
-/** @type {import('tailwindcss').Config} */
-export default withUt({
-  content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx}"],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-});
