+++ examples/minimal-express-react/README.md
@@ -1,0 +1,18 @@
+# Minimal Express example for UploadThing
+
+<a href="https://stackblitz.com/github/pingdotgg/uploadthing/tree/main/examples/minimal-express-react">
+  <img height="64" src="https://github.com/pingdotgg/uploadthing/assets/51714798/45907a4e-aa64-401a-afb3-b6c6df6eb71f" />
+</a>
+
+## QuickStart
+
+1. Grab an API key from the UploadThing dashboard:
+   https://uploadthing.com/dashboard
+2. `cp .env.example .env` and paste in your API key in the newly created `.env`
+   file
+3. `pnpm dev`
+4. Upload files!
+
+## Further reference
+
+Check out the docs at: https://docs.uploadthing.com/nextjs/pagedir
