+++ examples/minimal-express-react/client/index.html
@@ -4,7 +4,8 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + React + TS</title>
+    <script src="https://cdn.tailwindcss.com"></script>
+    <title>UploadThing + Express</title>
   </head>
   <body>
     <div id="root"></div>
