+++ examples/minimal-express-react/client/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@examples/minimal-express-react-client",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "dev": "vite --port 3001",
+    "build": "tsc && vite build",
+    "start": "vite preview"
+  },
+  "dependencies": {
+    "@uploadthing/react": "5.5.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-dropzone": "^14.2.3",
+    "uploadthing": "^5.5.0"
+  },
+  "devDependencies": {
+    "@types/react": "18.2.8",
+    "@types/react-dom": "18.2.4",
+    "@vitejs/plugin-react-swc": "^3.3.2",
+    "typescript": "^5.1.6",
+    "vite": "^4.4.9"
+  }
+}
