+++ examples/minimal-express-react/client/src/uploadthing.ts
@@ -1,0 +1,9 @@
+import { generateComponents } from "@uploadthing/react";
+import { generateReactHelpers } from "@uploadthing/react/hooks";
+
+import type { OurFileRouter } from "../../server";
+
+export const { UploadButton, UploadDropzone, Uploader } =
+  generateComponents<OurFileRouter>();
+
+export const { useUploadThing } = generateReactHelpers<OurFileRouter>();
