+++ examples/minimal-express-react/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@examples/minimal-express-react",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "workspaces": [
+    "client",
+    "server"
+  ],
+  "scripts": {
+    "dev": "concurrently \"npm run -w client dev\" \"npm run -w server dev\"",
+    "build": "npm run -w client build && npm run -w server build",
+    "start": "concurrently \"npm run -w client start\" \"npm run -w server start\""
+  },
+  "devDependencies": {
+    "concurrently": "^8.2.1",
+    "typescript": "^5.1.6"
+  }
+}
