+++ examples/minimal-express-react/server/package.json
@@ -1,10 +1,9 @@
 {
-  "name": "@example/express",
-  "version": "0.0.0",
-  "description": "",
-  "main": "index.js",
+  "name": "@example/minimal-express-react-server",
   "scripts": {
-    "dev": "tsx index.ts"
+    "build": "tsc",
+    "dev": "tsx index.ts",
+    "start": "node dist/index.js"
   },
   "keywords": [],
   "author": "",
@@ -13,12 +12,13 @@
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
-    "uploadthing": "*",
+    "uploadthing": "^5.5.0",
     "zod": "^3.21.4"
   },
   "devDependencies": {
     "@types/cors": "^2.8.13",
     "@types/express": "^4.17.17",
-    "tsx": "^3.12.7"
+    "tsx": "^3.12.7",
+    "typescript": "^5.1.6"
   }
 }
