+++ pnpm-workspace.yaml
@@ -2,5 +2,6 @@ packages:
   - "packages/config/*"
   - "packages/*"
   - "examples/*"
-  - "examples/express/*"
+  - "examples/*/*"
+  - "!examples/*/.next"
   - "docs"
